Fixed edited world not loading edited terrain data
while testing #162

More or less fixed issue #197

This is not a perfect fix becuase the output texture ignores the effect of shaders used by the raw image rendering materials. Therefore the resulting exported image doesn't look the same as the one on screen.

I tried implementing an alternative which involves rendering the actual map image into a secondary camera and then render that camera into a texture. But this is a complicated solution that requires quite a bit of work and testing. Id rather leave that for a future patch. Nevertheless I left some of the initial elements for it placed within the scene (disabled, of course)

fixed #199

partial fix to #200

this change attempts to reduce instances of miscalculations on drainage by first doing a "test" drainage expansion that finds all cells that need to be recalculated but without actually modifying them permanently. A second drainage expansion will only serve to modify the terrain (erosion and temperature carriage). The third and final drainage expansion will leave the final water accumulation values on each cell

Another semi-failed stab at fixing #200